(MAINT) Change `update-pdk-release-file.rb` to check multiple sources
This commit updates the `update-pdk-release-file.rb` script to check
both `nightlies.puppetlabs.com` and `apt.puppetlabs.com` for PDK
releases and use the actual package file timestamps to determine what
the "latest" release is between the two sources.

Once it has determined the latest release, the rest of the process is
mostly unchanged, it will update the `pdk-release.env` file with the
version and release type of the latest release. The only change is that
for non-nightly releases, the `PDK_DEB_URL` link will point to
`apt.puppetlabs.com` which should be a more stable source than nightlies
where packages can disappear after a while.