CAT-2195: Fix ci.yml with known issue fix
Add --template-ref, --template-url to fix known issue.  pdk actions may
fail with an error like `pdk (FATAL): Unable to find a
valid module template to use.`. This is a known issue probably caused
by a git dependency. The fix and the know issue are described here
<https://www.puppet.com/docs/pdk/3.x/pdk_known_issues.html#pdk-340---issue-with-utilising-bundled-templates>.
This commit simply adds the fix

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>